requirements.txt
# Core
python-dotenv
phi
groq

# Tools
yfinance
duckduckgo-search

# Playground / serving
uvicorn
fastapi

streamlit
langchain_core
langchain_community
langchain_ollama
pdfplumber